refactor(api): move UnmarshalParams to lib, lib.NewHTTPRequestHandler
Leveraging our new patterns to have lib begin to generate it's own request
handlers
